% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaQCAT.R
\name{MetaQCAT}
\alias{MetaQCAT}
\title{meta-analysis version of Quasi-Conditional Association Tests for microbiome data}
\usage{
MetaQCAT(
  OTU,
  X,
  X.index,
  meta.method = "FE-MetaQCAT",
  class.id = NULL,
  Tax = NULL,
  min.depth = 0,
  max.zero.prop = 0.9,
  n.perm = NULL,
  permute.strata = NULL,
  fdr.alpha = 0.05,
  n.threads = 1
)
}
\arguments{
\item{OTU}{a list of matrices contains OTU counts with each row corresponding to a sample and each column corresponding to an OTU or taxa. Column names, denoting the species names of each taxon, are mandatory.}

\item{X}{a list of matrices contains covariates, with each column pertaining to one variable. The number of elements in X and OTU must match.}

\item{X.index}{a vector indicates the columns in X for the covariate(s) of interest.}

\item{meta.method}{The meta-analysis method to be used. Including composition data based fixed effect test "FE-MetaQCAT", presence-absence data based fixed effect test "FE-MetaQCAT*" and the Cauchy-combined p-value of the two fix effect tests denoted as "FE-MetaQCAT-O",
composition data based random effect test "RE-MetaQCAT", presence-absence data based random effect test "RE-MetaQCAT*"  and the Cauchy-combined p-value of all the random effect tests denoted as "RE-MetaQCAT-O".}

\item{class.id}{a list of factors represents the group index to which each sample belongs if samples are correlated. The default is null, but this must be provided if samples are correlated.}

\item{Tax}{a matrix defines the taxonomy ranks , with each row corresponding to an OTU or taxa and each column to a rank (starting from the higher taxonomic level). Row names are mandatory and should be consistent with the column names of the OTU table. Column names should be formatted as "Rank1", "Rank2", etc.
If provided, tests will be performed for lineages based on the taxonomic rank. The output includes P-values for all lineages; a list of significant lineages controlling the false discovery rate (based on the resampling p-value if a resampling test is performed).
If not provided, a single test will be performed with all OTUs, and one p-value will be output.}

\item{min.depth}{keep samples with depths >= min.depth.}

\item{max.zero.prop}{keep taxa with percentage of zero counts <= max.zero.prop.}

\item{n.perm}{perform asymptotic test if n.perm is null, otherwise, perform permutation tests using the specified number of resamplings. If using random effect meta-analysis methods, then a resampling test must be performed}

\item{permute.strata}{a specification for how to perform resampling test for correlated samples. The default is null. The shuffle will be held within each group and among groups if the value is "within" and "between", respectively.}

\item{fdr.alpha}{false discovery rate for multiple tests on the lineages.}

\item{n.threads}{number of cores for parallel computation, with default being one.}
}
\value{
If Tax is null, only the global p-value for all OTUs will be returned. If Tax is provided, the output will include:
\item{lineage.pval}{p-values for all lineages. By default ( meta.method = "FE-MetaQCAT", n.perm = NULL ), only the asymptotic test will be performed.}
\item{sig.lineage}{a vector of significant lineages.}
\item{lineage.coef}{the estimated coefficent for each taxon, including the estimated parameter and the its variance. If provided method is composition data based, the taxon regarded as the reference will also be returned.}
}
\description{
meta-analysis version of Quasi-Conditional Association Tests for microbiome data
}
\references{
Tang ZZ, Chen G, Alekseyenko AV, Li H. (2017) A general framework for association analysis of microbial communities on a taxonomic tree.
\emph{Bioinformatics}
\doi{10.1093/bioinformatics/btw804}.

Lee S, Teslovich TM, Boehnke M, Lin X. (2013) General framework for meta-analysis of rare variants in sequencing association studies. Am J Hum Genet.
\emph{Am J Hum Genet}
\doi{10.1016/j.ajhg.2013.05.010}.

Benjamini, Yoav, and Yosef Hochberg.(1995) Controlling the False Discovery Rate: A Practical and Powerful Approach to Multiple Testing.
\emph{Journal of the Royal Statistical Society. Series B}
}
